Make sure correct value for byte[] is set in tag
The setting of the byte[] value in TagConventions was previously
overridden by the tags original value, when using BuildBy(...),
causing the submit of a department to fail. This is because
TagConventions relies on ToString() when setting the value.

Modify original tag's value to make sure new tag's value gets set
correctly.